OPCT-311: generate check rules documentation automatically (#129)
This change introduces a command to extract the available check
rules (SLOs) from the existing OPCT version and update the respective
documentation in markdown format.

This change ensure all available rules defined in the CLI, processed by
report, are well documented helping the reviewer to take actions
while reviewing the archive.

The bug is impacting partners when running self-review.