OPCT-270: Added arm64 instructions and multi-arch builds (#93)
Creating ARM64 support on OPCT validation environment, it is required to
build/mirror multi-arch (amd64 and arm64) the following components:
- sonobuoy aggregator server/worker image
- plugin openshift-tests image (and dependencies)
  - tools (base image) and dependencies (oc, jq, etc)
  - must-gather-monitoring image

The build for those components are implemented in the PR
https://github.com/redhat-openshift-ecosystem/provider-certification-plugins/pull/51

This PR bumps the Plugins image to use multi-arch payloads starting the
next release (v0.5.0-alpha.3). The documentation is also provided with
steps to create new builds and to quickly test in an ARM cluster, OCP on
AWS with IPI.

Note: I am not updating `mkdocs.html` to prevent indexing new docs in
the current documentation release. cc
[OPCT-251](https://issues.redhat.com/browse/OPCT-251)

https://issues.redhat.com/browse/OPCT-270

---------

Co-authored-by: Richard Vanderpool <49568690+rvanderp3@users.noreply.github.com>