OPCT-263: docs: v0.5 - added CLI reference (#148)
**What this PR does / why we need it**:

Introducing CLI reference guides to enhance the `opct`
documentation/options for users not applying for VCSP. This will include
a summary of each option and it's output.

It's not currently indexed by `mkdocs.yaml`, the changes is added in the
PR https://github.com/redhat-openshift-ecosystem/opct/pull/124 when
final v0.5 documentation will be released.

_Isolating changes from
https://github.com/redhat-openshift-ecosystem/opct/pull/124_

**Which issue(s) this PR fixes**:

https://issues.redhat.com/browse/OPCT-263

**Checklist**
- [x] Subject and description added to both, commit and PR.
- [x] Relevant issues have been referenced.
- [x] This change includes docs. 
- [ ] This change includes unit tests.