OPCT-350: introduce check/slo to evaluate valid install/platform for VCSP (#171)
**What this PR does / why we need it**:

This change introduces a new check to evaluate if the cluster has
platform type and install method required by applying a ticket to the
Red Hat Connect Portal for a OpenShift validation program (VCSP).

This is in conformance with the documentation and signalize
partners/reviewers that the archive have been executed in a cluster with
invalid configuration.


**Which issue(s) this PR fixes**

https://issues.redhat.com/browse/OPCT-350

**Checklist**
- [x] Subject and description added to both, commit and PR.
- [x] Relevant issues have been referenced.
- [x] This change includes docs. 
- [x] This change includes unit tests.