OPCT-326: bump to v0.6.0-alpha.2 to collect kube-burner data (#169)
**What this PR does / why we need it**:

This change introduces kube-burner performance data collection.

kube-burner is a standard tool to collect performance and scale tests at
OpenShift, those profiles would be the basic one to evaluate the current
cluster state with small size.

It's not the goal to extract insights, we need first to collect data
points to a controlled environment before proceeding with any
evaluation/report.

Blocked by:
- [x]
https://github.com/redhat-openshift-ecosystem/provider-certification-plugins/pull/67


**Which issue(s) this PR fixes** *(optional, use `fixes
#<issue_number>(, fixes #<issue_number>, ...)` format, where
issue_number might be a GitHub issue, or a Jira story*:
Fixes #

https://issues.redhat.com/browse/OPCT-328



**Checklist**
- [ ] Subject and description added to both, commit and PR.
- [x] Relevant issues have been referenced.
- [ ] This change includes docs. 
- [ ] This change includes unit tests.