OPCT-308: Introduce check to validate cluster install time (#170)
**What this PR does / why we need it**:

This change is updating the Checks/SLOs, introducing the check
OPCT-032E to validate the total time since the cluster was installed
and conformance tests had been executed.

This is a common issue when partners installing a cluster, runs many
conformance tests executions, or just run the final conformance in old
clusters which logs and metrics have been contaminated with
non-conformance tests executions, making the framework to raise
false-positive signals.

This check is satisfying the diagram introduced in:
https://github.com/redhat-openshift-ecosystem/opct/pull/168

https://redhat-openshift-ecosystem.github.io/opct/guides/cluster-validation/#process-overview

which guides the partner to open a ticket when executed in a fresh
cluster.

**Which issue(s) this PR fixes** :

Fixes #
https://issues.redhat.com/browse/OPCT-308

**Checklist**
- [x] Subject and description added to both, commit and PR.
- [x] Relevant issues have been referenced.
- [x] This change includes docs. 
- [ ] This change includes unit tests.